Update readme.md
This is my first attempt at creating a scene in WebGL without any tutorial or guidance. I did download a boilerplate webpack from Bruno Simon's Threejs Journey as a great starting point. 

This scene focuses on a fictional alien landscape. At the moment I'm using a Neptune texture for the planet as well as a generic snow texture for the landscape.